Default theme asset path can be overridden in parts
./bin/codo has a new option '--theme-local'. Setting this option to the
name of a directory (default: './codo') will make the Templater of the
default theme look therefor assets. This way, the CSS can be overridden
without creating a full-blown theme.

To make customization easier, color and font declarations are
centralized in a file `vars.styl`. This is not yet complete.

Example:
```
$> tree ./codo-override
  └── assets
      └── stylesheets
          ├── alphabetical_index.styl
          ├── application.styl
          ├── base.styl
          ├── class.styl
          ├── footer.styl
          ├── header.styl
          ├── lists.styl
          ├── toc.styl
          ├── vars.styl
          └── vendor
              └── highlight.css
$> codo --theme-local ./codo-override SomeFile.coffee
```